feat(session) move login methods from store to service
Refers to #4
