--- src/app/core/state/session/index.ts
@@ -1,4 +1,0 @@
-export * from './session.query';
-export * from './session.store';
-export * from './session.service';
-export * from './session.interface';
