+++ src/app/core/state/session/session.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { tap } from 'rxjs/operators';
 import { SessionStore } from './session.store';
-import { SessionApiService } from '../../services/session/session-api.service';
+import { SessionApiService } from '../../api/session/session-api.service';
 
 @Injectable({ providedIn: 'root' })
 export class SessionService {
@@ -9,8 +9,16 @@ export class SessionService {
 
   login() {
     return this.sessionApiService.postLogin().pipe(
-      tap((session) => {
-        this.sessionStore.login(session);
+      tap(session => {
+        this.sessionStore.update({ session });
+      })
+    );
+  }
+
+  logout() {
+    return this.sessionApiService.postLogout().pipe(
+      tap(() => {
+        this.sessionStore.update({ session: { accessToken: undefined, tokenType: undefined } });
       })
     );
   }
